+++ package.json
@@ -18,10 +18,6 @@
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/react-switch": "^1.1.3",
-    "@react-pdf-viewer/core": "3.12.0",
-    "@react-pdf-viewer/full-screen": "^3.12.0",
-    "@react-pdf-viewer/get-file": "^3.12.0",
-    "@react-pdf-viewer/zoom": "^3.12.0",
     "@t3-oss/env-nextjs": "^0.10.1",
     "@types/mongoose": "^5.11.97",
     "@ungap/with-resolvers": "^0.1.0",
