+++ src/app/api/ai-upload/route.ts
@@ -19,8 +19,7 @@ import { PaperAdmin } from "@/db/papers";
 import axios from "axios";
 import processAndAnalyze from "@/util/gemini";
 import Fuse from "fuse.js";
-// import processAndAnalyze from "./mistral";
-// TODO: REMOVE THUMBNAIL FROM admin-buffer DB
+
 cloudinary.v2.config({
   cloud_name: process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
